% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_columns.R
\name{all_columns}
\alias{all_columns}
\title{All columns for the projects, files, samples, and bundles endpoints}
\usage{
all_columns(type = NULL)
}
\arguments{
\item{type}{character() type of all columns wanted; one of either "projects",
"files", "samples", or "bundles"}
}
\value{
character() vector of all columns formatted like "hits\link{*}..."
}
\description{
\code{all_columns()} queries the API for all available columns for
either projects, files, samples, or bundles
}
\examples{
all_columns("projects")

}
