% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_function.R
\name{anno_function}
\alias{anno_function}
\title{anno_function}
\usage{
anno_function(data = data, sig_thres = 0.2, min_cell = 50)
}
\arguments{
\item{data}{your Seurat object(support V4)}

\item{sig_thres}{threshold for signature scores}

\item{min_cell}{cell types contain a reasonable number of cells}
}
\value{
Seurat object with annotations
}
\description{
anno_function
}
