% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_function.R
\name{anno_immune}
\alias{anno_immune}
\title{anno_immune}
\usage{
anno_immune(
  datafilt,
  scGate_DB = scGate_DB,
  non_epi = non_epi,
  min_cell = 100,
  MAGIC_impute = FALSE,
  organism = "human",
  ncore = 30
)
}
\arguments{
\item{organism}{human or mouse}

\item{ncore}{}
}
\description{
anno_immune
}
