% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_function.R
\name{anno_tumor}
\alias{anno_tumor}
\title{Title}
\usage{
anno_tumor(
  datafilt = datafilt,
  cnadata = cnadata,
  thres_sig = 0.005,
  thres_cor = 0.5,
  ncore = 30,
  isFilter = FALSE
)
}
\arguments{
\item{datafilt}{your Seurat object(support V4)}

\item{ncore}{number of core}

\item{isFilter}{filter Endothelial/Stomal cells}
}
\description{
Title
}
