% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_function.R
\name{auto_immune}
\alias{auto_immune}
\title{auto_immune}
\usage{
auto_immune(
  datafilt,
  scGate_DB = scGate_DB,
  non_epi = non_epi,
  min_cell = 100,
  MAGIC_impute = FALSE,
  ncore = 30
)
}
\arguments{
\item{ncore}{}
}
\description{
auto_immune
}
