% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_function.R
\name{autoumap}
\alias{autoumap}
\title{autoumap}
\usage{
autoumap(datafilt, nfeatures = 2000, ndim = 15, neigh = 20, dist = 1)
}
\arguments{
\item{datafilt}{your Seurat object(support V4)}

\item{nfeatures}{Number of features to select as top variable features}

\item{ndim}{Which dimensions to use as input features}

\item{neigh}{This determines the number of neighboring points used in local approximations of manifold structure.}

\item{dist}{This controls how tightly the embedding is allowed compress points together.}
}
\value{
Seurat object with umap
}
\description{
autoumap
}
