% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_function.R
\name{cna_heatmap}
\alias{cna_heatmap}
\title{Title}
\usage{
cna_heatmap(
  cnadata = cnadata,
  thres_sig = 0.005,
  thres_cor = 0.5,
  max_cell = 15000
)
}
\arguments{
\item{max_cell}{}
}
\description{
Title
}
