% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_query_filter.R
\name{construct_query_filter}
\alias{construct_query_filter}
\title{Query filter construction function}
\usage{
construct_query_filter(
  r_filter = list(),
  base_url = "https://service.dev.singlecell.gi.ucsc.edu",
  endpoint = "/index/projects",
  params = list()
)
}
\arguments{
\item{r_filter}{list() query filter formatted as a nested list}

\item{base_url}{character(1)}

\item{endpoint}{character(1)}

\item{params}{list() additional query parameters formatted as a nested list}
}
\value{
character(1) url for API call
}
\description{
function for constructing the query url; defaults to querying
Human Cell Atlas API
}
