% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_url.R
\name{construct_url}
\alias{construct_url}
\title{Title}
\usage{
construct_url()
}
\value{

}
\description{
Title
}
\examples{
construct_url(
parameters = parameters(),
endpoint = "/index/projects",
base_url = "https://service.dev.singlecell.gi.ucsc.edu"
)
}
