% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.R
\name{.details}
\alias{.details}
\title{Single Entity Details}
\usage{
.details(
  uuid = character(),
  catalog = c("dcp2", "it2", "dcp1", "it1"),
  view = c("projects", "files", "samples", "bundles")
)
}
\arguments{
\item{uuid}{character() unique *_id}

\item{catalog}{character() catalog entity belongs to}

\item{view}{character() type of entity i.e. project, file, sample, or bundle}
}
\value{
list-of-lists containing relevant details about the project, file,
sample, or bundle
}
\description{
Single Entity Details
}
