% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filters}
\alias{filters}
\alias{length.filters}
\alias{print.filters}
\title{FIXME}
\usage{
filters(...)

\method{length}{filters}(x)

\method{print}{filters}(x, ...)
}
\arguments{
\item{...}{named arguments, each of which is a \code{list()} specifying
... FIXME}

\item{x}{for \code{length()} and \code{print()}, an object of class \code{filters}.}
}
\value{
\code{filters()} returns a \code{filters} object representing
validated filters in a format suitable for use in \code{projects()}
and related functions.
}
\description{
Functions for the creat
}
\examples{
filters()

filters(
    organ = list(is = "pancreas")
)

filters(
    organ = list(is = "pancreas"),
    genusSpecies = list(is = "Homo sapiens")
)

filters(
    fileFormat = list(is = c("fastq", "fastq.gz"))
)

}
