% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_api_call.R
\name{hca_api_call}
\alias{hca_api_call}
\title{Human Cell Atlas API call}
\usage{
hca_api_call(url = "")
}
\arguments{
\item{url}{character(1) url string for API call}
}
\value{
list() response headers and content from call, error message outpt if
needed
}
\description{
function makes a GET request to the Human Cell Atlas API using
the provided url string
}
