% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_base_call.R
\name{hca_base_call}
\alias{hca_base_call}
\title{Full Human Cell Atlas API call function}
\usage{
hca_base_call(
  r_filter = list(),
  base_url = "https://service.dev.singlecell.gi.ucsc.edu",
  endpoint = "/index/projects",
  params = list()
)
}
\arguments{
\item{r_filter}{list() query filter formatted as a nested list}

\item{base_url}{character(1)}

\item{endpoint}{character(1)}

\item{params}{list() additional query parameters formatted as a nested list}
}
\value{
list() response headers and content from call, error message output
if needed
}
\description{
function is composed of the construct_query_filter() and
hca_api_call() functions, taking in API call parameters and returning a
response
}
