% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hca_view.R
\name{hca_view}
\alias{hca_view}
\title{View and select table rows interactively}
\usage{
hca_view(tbl)
}
\arguments{
\item{tbl}{a 'tibble' of \code{projects()}, \code{samples()}, \code{bundles()}, or
\code{files()}.}
}
\value{
\code{hca_view()} returns a tibble filtered to reflect the rows
selected in the interface.
}
\description{
View and select table rows interactively
}
\examples{
if (interactive()) {
    p <- projects(size = 100)
    p1 <- hca_view(p)  # interactive table browser; filtered results
}
}
