% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_function.R
\name{immune}
\alias{immune}
\title{immune cell annotation}
\usage{
immune(
  datafilt,
  scGate_DB = scGate_DB,
  non_epi = non_epi,
  min_cell = 100,
  organism = "human",
  ncore = 30
)
}
\arguments{
\item{datafilt}{your Seurat object(support V4)}

\item{scGate_DB}{scGate gene list}

\item{non_epi}{epi marker genes}

\item{min_cell}{cell types contain a reasonable number of cells}

\item{organism}{human or mouse}

\item{ncore}{number of cores}
}
\value{
Seurat object with immune cell annotation
}
\description{
immune cell annotation
}
