% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_function.R
\name{infer_tumor}
\alias{infer_tumor}
\title{infer_tumor}
\usage{
infer_tumor(
  datafilt = datafilt,
  cnadata = cnadata,
  thres_sig = 0.005,
  thres_cor = 0.5,
  organism = "human",
  ncore = 30,
  isFilter = FALSE
)
}
\arguments{
\item{datafilt}{your Seurat object(support V4)}

\item{organism}{human or mouse}

\item{ncore}{number of core}

\item{isFilter}{filter Endothelial/Stomal cells}
}
\description{
infer_tumor
}
