% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_function.R
\name{infertumor}
\alias{infertumor}
\title{infertumor}
\usage{
infertumor(
  datafilt = datafilt,
  scGate_DB = scGate_DB,
  thres_sig = 0.005,
  thres_cor = 0.5,
  isFilter = FALSE,
  ncore = 30
)
}
\arguments{
\item{ncore}{}
}
\description{
infertumor
}
