% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_function.R
\name{integrate}
\alias{integrate}
\title{integrate}
\usage{
integrate(
  dataimmu = dataimmu,
  datacanc = datacanc,
  min_tumor = 50,
  rm_doublet = FALSE,
  prop_doublet = 0.075
)
}
\arguments{
\item{prop_doublet}{}
}
\description{
integrate
}
