% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom_annotation.R
\name{loom_annotation}
\alias{loom_annotation}
\title{HCA loom file annotation}
\usage{
loom_annotation(loom_file_path = NULL)
}
\arguments{
\item{loom_file_path}{character() file path of loom file on user's system.}
}
\value{
SummarizedExperiment object
}
\description{
\code{loom_annotation()} takes the file path location of a .loom file
for which additional data will be extracted from the appropriate manifest.
The .loom file will be imported as a SummarizedExperiment object, and the
additional manifest information will be added to the object for return.
}
\seealso{
\code{manifest()} and related functions for working with data returned
from the \verb{*/manifest/*} HCA API endpoints.
}
