% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Title}
\usage{
parameters()
}
\value{

}
\description{
Title
}
\examples{
parameters(
filter = filters(),
catalog = "dcp2",
size = 10,
sort = "projectTitle",
order = "asc"
)
}
