% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibbles.R
\name{next_tbl}
\alias{next_tbl}
\alias{prev_tbl}
\title{Tibble Functions}
\usage{
next_tbl(curr_tbl)

prev_tbl(curr_tbl)
}
\arguments{
\item{curr_tbl}{tibble current tibble for which we want the previous page}
}
\value{
tibble

tibble
}
\description{
Tibble Functions
}
