% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno_function.R
\name{tumor}
\alias{tumor}
\title{tumor annotation}
\usage{
tumor(datafilt = datafilt, scGate_DB = scGate_DB, ncore = 30)
}
\arguments{
\item{datafilt}{your Seurat object(support V4)}

\item{scGate_DB}{scGate gene list}

\item{ncore}{numbers of cores}
}
\value{
Seurat object
}
\description{
tumor annotation
}
