% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundles.R
\name{bundles}
\alias{bundles}
\alias{bundles_facets}
\alias{bundles_default_columns}
\alias{bundles_detail}
\title{HCA File Querying}
\usage{
bundles(
  filters = NULL,
  size = 1000L,
  sort = "projectTitle",
  order = c("asc", "desc"),
  catalog = c("dcp2", "it2", "dcp1", "it1"),
  as = c("tibble", "lol"),
  columns = bundles_default_columns("character")
)

bundles_facets(facet = character(), catalog = c("dcp2", "it2", "dcp1", "it1"))

bundles_default_columns(as = c("tibble", "character"))

bundles_detail(uuid, catalog = c("dcp2", "it2", "dcp1", "it1"))
}
\arguments{
\item{filters}{filter object created by \code{filters()}, or \code{NULL}
(default; all projects).}

\item{size}{integer(1) maximum number of results to return;
default: all projects matching \code{filter}. The default (10000) is
meant to be large enough to return all results.}

\item{sort}{character(1) project facet (see \code{facet_options()}) to
sort result; default: \code{"projectTitle"}.}

\item{order}{character(1) sort order. One of \code{"asc"} (ascending) or
\code{"desc"} (descending).}

\item{catalog}{character() catalog to use for query.}

\item{as}{character(1) return format. Default: \code{"tibble"}, a tibble
summarizing essential elements of HCA projects. \code{"lol"}: a
list-of-lists containing detailed project information}

\item{columns}{named character() indicating the paths to be used
for parsing the 'lol' returned from the HCA to a tibble. The
names of \code{columns} are used as column names in the returned
tibble.}

\item{facet}{character() of valid facet names. Summary results (see
'Value', below) are returned when missing or length greater
than 1; details are returned when a single facet is specified.}

\item{uuid}{character() unique identifier (e.g., \code{projectId}) of
the object.}
}
\value{
When \code{as = "tibble"}, \code{bundles()} returns a tibble with each
row representing a file for one of the specified HCA project,
and columns summarizing the file.  Each \code{.hit} is a single
result; ....\preformatted{When `as = "lol"`, `bundles()` returns a list-of-lists data
structure representing detailed information on each file
(`hit`).
}

\code{bundles_detail()} returns a list-of-lists containing
relevant details about the bundle
}
\description{
\code{bundles()} takes a list of user provided project titles
to be used to query the HCA API for information about available bundles.

\code{bundles_detail()} takes a unique bundle_id and catalog for
the bundle, and returns details about the specified bundle as a
list-of-lists
}
\examples{
bundles(filters = filters(
    projectTitle = list(
        is = c("Tabula Muris: Transcriptomic characterization of 20 organs
        and tissues from Mus musculus at single cell resolution")
   )
))
bundles_facets()

bundles_detail(
    uuid = "00aa6c53-71b5-4c12-98c4-54eb8173ffa5",
    catalog = "dcp2"
)

}
\seealso{
\code{lol_find()} and \code{lol_lfind()} for working with
list-of-lists data structures.
}
