% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogs.R
\name{catalogs}
\alias{catalogs}
\title{Title}
\usage{
catalogs()
}
\value{
character() vector of available catalogs
}
\description{
\code{catalogs()} queries the API for all available project catalogs
}
\examples{
catalogs()
}
