% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files}
\alias{files}
\alias{download_files}
\title{HCA File Querying}
\usage{
files(
  filters = NULL,
  size = 1000L,
  sort = "projectTitle",
  order = c("asc", "desc"),
  catalog = c("dcp2", "it2", "dcp1", "it1"),
  as = c("tibble", "lol")
)

download_files(files_tib = files(), destination = tempdir())
}
\arguments{
\item{filters}{filter object created by \code{filters()}, or \code{NULL}
(default; all files).}

\item{size}{integer(1) maximum number of results to return;
default: all files matching \code{filter}. The default (10000) is
meant to be large enough to return all results.}

\item{sort}{character(1) project facet (see \code{facet_options()}) to
sort result; default: \code{"projectTitle"}.}

\item{order}{character(1) sort order. One of \code{"asc"} (ascending) or
\code{"desc"} (descending).}

\item{catalog}{character(1) source of data. Default: \code{"dcp2"},
version 2 of the HCA Data Coordinating Platform.}

\item{as}{character(1) return format. Default: \code{"tibble"}, a tibble
summarizing essential elements of HCA files. \code{"lol"}: a
list-of-lists containing detailed file information.}

\item{files_tib}{tibble of files (result of \code{files()})}

\item{destination}{}
}
\value{
When \code{as = "tibble"}, \code{files()} returns a tibble with each
row representing a file for one of the specified HCA project,
and columns summarizing the file.  Each \code{.hit} is a single
result; ....\preformatted{When `as = "lol"`, `files()` returns a list-of-lists data
structure representing detailed information on each file
(`hit`).
}

file_destinations vector of file destinations
}
\description{
\code{files()} takes a list of user provided project titles
to be used to query the HCA API for information about available files.

\code{download_files()} takes a tibble of files and a directory
location as arguments to download the files of the tibble into the specified
directory.
}
\examples{
files(filters = filters(
    projectTitle = list(
        is = c("Tabula Muris: Transcriptomic characterization of 20 organs
        and tissues from Mus musculus at single cell resolution")
   )
))
download_files(file_tib = files(filter = filters(
projectId = list(is = projectId),
fileFormat = list(is = "loom"))), destination = tempdir())
}
\seealso{
\code{lol_find()} and \code{lol_lfind()} for working with
list-of-lists data structures.
}
