% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files}
\alias{files}
\title{HCA File Querying}
\usage{
files(
  filters = NULL,
  size = 1000L,
  sort = "projectTitle",
  order = c("asc", "desc"),
  catalog = c("dcp2", "it2", "dcp1", "it1")
)
}
\arguments{
\item{filters}{filter object created by \code{filters()}, or \code{NULL}
(default; all projects).}

\item{size}{integer(1) maximum number of results to return;
default: all projects matching \code{filter}. The default (10000) is
meant to be large enough to return all results.}

\item{sort}{character(1) project facet (see \code{facet_options()}) to
sort result; default: \code{"projectTitle"}.}

\item{order}{character(1) sort order. One of \code{"asc"} (ascending) or
\code{"desc"} (descending).}

\item{catalog}{character(1) source of data. Default: \code{"dcp2"},
version 2 of the HCA Data Coordinating Platform.}
}
\value{
\code{files()} returns a tibble with each row representing a file for one
of the specified HCA project, and columns summarizing the file.
Each \code{.hit} is a single result; ....
}
\description{
\code{files()} takes a list of user provided project titles
to be used to query the HCA API for information about available files.
}
\examples{
files(filters = filters(
    projectTitle = list(
        is = c("Tabula Muris: Transcriptomic characterization of 20 organs
        and tissues from Mus musculus at single cell resolution")
   )
))
}
