% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{files}
\alias{files}
\alias{files_default_columns}
\alias{files_download}
\alias{files_facets}
\alias{files_detail}
\title{HCA File Querying}
\usage{
files(
  filters = NULL,
  size = 1000L,
  sort = "projectTitle",
  order = c("asc", "desc"),
  catalog = c("dcp2", "it2", "dcp1", "it1"),
  as = c("tibble", "lol"),
  columns = files_default_columns("character")
)

files_default_columns(as = c("tibble", "character"))

files_download(tbl, destination = tempdir())

files_facets(facet = character(), catalog = c("dcp2", "it2", "dcp1", "it1"))

files_detail(uuid, catalog = c("dcp2", "it2", "dcp1", "it1"))
}
\arguments{
\item{filters}{filter object created by \code{filters()}, or \code{NULL}
(default; all projects).}

\item{size}{integer(1) maximum number of results to return;
default: all projects matching \code{filter}. The default (10000) is
meant to be large enough to return all results.}

\item{sort}{character(1) project facet (see \code{facet_options()}) to
sort result; default: \code{"projectTitle"}.}

\item{order}{character(1) sort order. One of \code{"asc"} (ascending) or
\code{"desc"} (descending).}

\item{catalog}{character() catalog to use for query.}

\item{as}{character(1) return format. Default: \code{"tibble"}, a tibble
summarizing essential elements of HCA projects. \code{"lol"}: a
list-of-lists containing detailed project information}

\item{columns}{named character() indicating the paths to be used
for parsing the 'lol' returned from the HCA to a tibble. The
names of \code{columns} are used as column names in the returned
tibble.}

\item{tbl}{tibble of files (result of \code{files()})}

\item{destination}{character() vector name of temporary directory to use
for file downloads}

\item{facet}{character() of valid facet names. Summary results (see
'Value', below) are returned when missing or length greater
than 1; details are returned when a single facet is specified.}

\item{uuid}{character() unique identifier (e.g., \code{projectId}) of
the object.}
}
\value{
When \code{as = "tibble"}, \code{files()} returns a tibble with each
row representing a file for one of the specified HCA project,
and columns summarizing the file.  Each \code{.hit} is a single
result; ....\preformatted{When `as = "lol"`, `files()` returns a list-of-lists data
structure representing detailed information on each file
(`hit`).
}

file_destinations vector of file destinations

\code{files_detail()} returns a list-of-lists containing
relevant details about the file.
}
\description{
\code{files()} takes a list of user provided project titles
to be used to query the HCA API for information about available files.

\code{files_download()} takes a tibble of files and a directory
location as arguments to download the files of the tibble into the specified
directory.

\code{files_detail()} takes a unique file_id and catalog for
the file, and returns details about the specified file as a
list-of-lists
}
\examples{
files(filters = filters(
    projectTitle = list(
        is = c("Tabula Muris: Transcriptomic characterization of 20 organs
        and tissues from Mus musculus at single cell resolution")
   )
))
files_filter <- filters(
    projectId = list(is = "cddab57b-6868-4be4-806f-395ed9dd635a"),
    fileFormat = list(is = "loom")
)
files_tbl <- files(filter = files_filter)
files_download(files_tbl, destination = tempdir())
files_facets()
files_facets("fileFormat")

files_detail(
    uuid = "bb4185da-c3da-4bb3-ab51-30aafbb60a0d",
    catalog = "dcp2"
)

}
\seealso{
\code{lol_find()} and \code{lol_lfind()} for working with
list-of-lists data structures.
}
