% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{.filters_validate}
\alias{.filters_validate}
\alias{filters}
\alias{length.filters}
\alias{print.filters}
\title{HCA Filter Construction}
\usage{
.filters_validate(x)

filters(...)

\method{length}{filters}(x)

\method{print}{filters}(x, ...)
}
\arguments{
\item{x}{for \code{length()} and \code{print()}, an object of class \code{filters}.}

\item{...}{named arguments, each of which is a \code{list()} specifying a query
facet and its corresponding value to be used in the query}
}
\value{
\code{filters()} returns a \code{filters} object representing
validated filters in a format suitable for use in \code{projects()}
and related functions.
}
\description{
This set of functions takes user input to be used as query
filters, ideally in the form of nested lists. This input is then validated,
reformatted to JSON, and encoded into a properly formatted URL.
}
\examples{
filters()

filters(
    organ = list(is = "pancreas")
)

filters(
    organ = list(is = "pancreas"),
    genusSpecies = list(is = "Homo sapiens")
)

filters(
    fileFormat = list(is = c("fastq", "fastq.gz"))
)

}
