% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest_url_generator}
\alias{manifest_url_generator}
\alias{manifest_uuid_constructor}
\title{Internal functions used by \code{manifest()}}
\usage{
manifest_url_generator(manifest_filter, catalog)

manifest_uuid_constructor(manifest_filter, catalog)
}
\arguments{
\item{manifest_filter}{hca filter object}

\item{catalog}{character() name of catalog}
}
\value{
\code{manifest_url_generator()} returns a url string to be used
in an API call

\code{manifest_uuid_constructor()} returns a url string to be
used in an API call
}
\description{
\code{manifest_url_generator()} takes a filter object with
criteria for the query and the catalog to search within.

\code{manifest_uuid_constructor()} takes a filter object
with criteria for the query and the catalog to search within.
}
\examples{
hca:::manifest_url_generator(filters(), catalogs()[1])
hca:::manifest_uuid_constructor(filters(), catalogs()[1])
}
\keyword{internal}
