% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimus_loom_annotation.R
\name{optimus_loom_annotation}
\alias{optimus_loom_annotation}
\title{HCA loom file annotation}
\usage{
optimus_loom_annotation(loom_file_path = NULL, catalog = NULL)
}
\arguments{
\item{loom_file_path}{character() file path of loom file on user's system.}

\item{catalog}{character() HCA catalog from which the .loom file originated.}
}
\value{
SingleCellExperiment object
}
\description{
\code{optimus_loom_annotation()} takes the file path location of a .loom file
for which additional data will be extracted from the appropriate manifest.
The .loom file will be imported as a LoomExperiment object, and the
additional manifest information will be added to the object for return.
}
\seealso{
\code{manifest()} and related functions for working with data returned
from the \verb{*/manifest/*} HCA API endpoints.
}
