% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects}
\alias{projects}
\title{FIXME}
\usage{
projects(filters = filters(), n = Inf, FIXME)
}
\arguments{
\item{filters}{FIXME}

\item{n}{integer(1) maximum number of results to return; default:
all projects mathcing \code{filter}.}
}
\value{
FIXME
}
\description{
FIXME
}
\examples{
FIXME

}
