% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{.parameters_validate}
\alias{.parameters_validate}
\alias{projects}
\title{HCA Project Querying}
\usage{
.parameters_validate(x)

projects(filters = filters(), n = Inf)
}
\arguments{
\item{filters}{filter object as defined in this package}

\item{n}{integer(1) maximum number of results to return; default:
all projects matching \code{filter}.}
}
\value{
\code{projects()} returns a tibble with each row representing an
HCA project, and columns summarizing the project.
}
\description{
\code{projects()} takes user input to be used to query the HCA API.
This function makes use of several other functions within this package.
}
\examples{
projects(filters(), n=20)

}
