% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R, R/tibbles.R
\name{projects}
\alias{projects}
\alias{projects_terms}
\alias{tibble_default_columns}
\title{HCA Project Querying}
\usage{
projects(
  filters = NULL,
  size = 1000L,
  sort = "projectTitle",
  order = c("asc", "desc"),
  catalog = c("dcp2", "it2", "dcp1", "it1"),
  as = c("tibble", "lol"),
  columns = tibble_default_columns("projects", "character")
)

projects_terms(facet = character(), catalog = c("dcp2", "it2", "dcp1", "it1"))

tibble_default_columns(
  view = c("projects", "files", "samples", "bundles"),
  as = c("tibble", "character")
)
}
\arguments{
\item{filters}{filter object created by \code{filters()}, or \code{NULL}
(default; all projects).}

\item{size}{integer(1) maximum number of results to return;
default: all projects matching \code{filter}. The default (10000) is
meant to be large enough to return all results.}

\item{sort}{character(1) project facet (see \code{facet_options()}) to
sort result; default: \code{"projectTitle"}.}

\item{order}{character(1) sort order. One of \code{"asc"} (ascending) or
\code{"desc"} (descending).}

\item{catalog}{character(1) source of data. Default: \code{"dcp2"},
version 2 of the HCA Data Coordinating Platform.}

\item{as}{character(1) return format. Default: \code{"tibble"}, a tibble
summarizing essential elements of HCA projects. \code{"lol"}: a
list-of-lists containing detailed project information}

\item{facet}{character() of valid facet names. Summary results (see
'Value', below) are returned when missing or length greater
than 1; details are returned when a single facet is specified.}

\item{view}{character(1) one of \code{"projects"}, \code{"files"},
\code{"samples"}, or \code{"bundles"}, corresponding to the name of the
function for which default values are desired.}
}
\value{
When \code{as = "tibble"}, \code{projects()} returns a tibble with
each row representing an HCA project, and columns summarizing
the project. Each \code{.hit} is a single result; a result may
contain several projects, as indexed by \code{.project}.\preformatted{When `as = "lol"`, `projects()` returns a list-of-lists data
structure representing detailed information on each project
(`hit`).
}

\code{projects_terms()} invoked with no \verb{facet=} argument returns a
tibble summarizing terms available as \code{projects()} return
values, and for use in filters. The tibble contains columns\preformatted{- `facet`: the name of the facet.
- `n_terms`: the number of distinct values the facet can take.
- `n_values`: the number of occurrences of the facet term in the
   entire catalog.

`projects_terms()` invoked with a scalar value for `facet=`
returns a tibble summarizing terms used in the facet, and the
number of occurrences of the term in the entire catalog.
}

\code{tibble_default_columns()} returns a tibble with column
\code{name} containing the column name used in the tibble returned
by \code{projects()}, \code{files()}, \code{samples()}, or \code{bundles()}, and
\code{path} the path (see \code{lol_hits()}) to the data in the
list-of-lists by the same functions when \code{as = "lol"}.
}
\description{
\code{projects()} takes user input to be used to query the
HCA API for information about available projects.

\code{projects_terms()} summarizes facets and terms used by
all records in the projects index.

\code{tibble_default_columns()} returns a tibble describing
the content of the tibble returned by \code{projects()}, \code{files()},
\code{samples()}, and `bundles().
}
\examples{
projects(filters())

projects_terms()
projects_terms("genusSpecies")

tibble_default_columns("projects")

}
\seealso{
\code{lol_find()} and \code{lol_lfind()} for working with
list-of-lists data structures.
}
