% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects}
\alias{projects}
\title{HCA Project Querying}
\usage{
projects(
  filters = NULL,
  size = 1000L,
  sort = "projectTitle",
  order = c("asc", "desc"),
  catalog = c("dcp2", "it2", "dcp1", "it1"),
  as = c("tibble", "lol")
)
}
\arguments{
\item{filters}{filter object created by \code{filters()}, or \code{NULL}
(default; all projects).}

\item{size}{integer(1) maximum number of results to return;
default: all projects matching \code{filter}. The default (10000) is
meant to be large enough to return all results.}

\item{sort}{character(1) project facet (see \code{facet_options()}) to
sort result; default: \code{"projectTitle"}.}

\item{order}{character(1) sort order. One of \code{"asc"} (ascending) or
\code{"desc"} (descending).}

\item{catalog}{character(1) source of data. Default: \code{"dcp2"},
version 2 of the HCA Data Coordinating Platform.}

\item{as}{character(1) return format. Default: \code{"tibble"}, a tibble
summarizing essential elements of HCA projects. \code{"lol"}: a
list-of-lists containing detailed project information}
}
\value{
When \code{as = "tibble"}, \code{projects()} returns a tibble with
each row representing an HCA project, and columns summarizing
the project. Each \code{.hit} is a single result; a result may
contain several projects, as indexed by \code{.project}.\preformatted{When `as = "lol"`, `projects()` returns a list-of-lists data
structure representing detailed information on each project
(`hit`).
}
}
\description{
\code{projects()} takes user input to be used to query the
HCA API for information about available projects.
}
\examples{
projects(filters())

}
\seealso{
\code{lol_find()} and \code{lol_lfind()} for working with
list-of-lists data structures.
}
