% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{projects}
\alias{projects}
\title{HCA Project Querying}
\usage{
projects(
  filters = filters(),
  size = 1000L,
  sort = "projectTitle",
  order = c("asc", "desc"),
  catalog = c("dcp2", "it2", "dcp1", "it1")
)
}
\arguments{
\item{filters}{filter object as defined in this package}

\item{size}{integer(1) maximum number of results to return;
default: all projects matching \code{filter}. The default (10000) is
meant to be large enough to return all results.}

\item{sort}{character(1) project facet (see \code{facet_options()}) to
sort result; default: \code{"projectTitle"}.}

\item{order}{character(1) sort order. One of \code{"asc"} (ascending) or
\code{"desc"} (descending).}

\item{catalog}{character(1) source of data. Default: \code{"dcp2"},
version 2 of the HCA Data Coordinating Platform.}
}
\value{
\code{projects()} returns a tibble with each row representing an
HCA project, and columns summarizing the project.
}
\description{
\code{projects()} takes user input to be used to query the
HCA API for information about available projects.
}
\examples{
projects()  # all projects

}
